backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pandas==2.0.2
pdfplumber==0.9.0
camelot-py==0.11.0
scikit-learn==1.2.2
python-multipart==0.0.6
numpy==1.24.3
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.1.0